refactor(App): App클래스가 담당하던 controller 기능 이전
코드 시작점 역할인 index.js 추가
App클래스 내 메서드 및 필드 RacingGameController클래스로 이전
