+++ src/index.js
@@ -1,0 +1,5 @@
+import App from './App.js';
+
+const app = new App();
+
+app.start();
